move files so github actions can find them
