+++ partners/actions/update-listing/mpapihelper.py
