+++ partners/actions/update-listing/newArtifact.json
